+++ client/src/components/RelatedProducts/Cards/Individualcard.jsx
