+++ client/src/components/RelatedProducts/Cards/sampleRelatedThumbnails.js
